requirements.txt
# Azure Functions
azure-functions>=1.18.0
azure-functions-worker>=1.0.0

# Core dependencies
beautifulsoup4==4.13.4
extract-msg==0.54.1
html2text==2025.4.15
pdfminer.six==20250506
pillow==11.3.0
pytesseract==0.3.13
python-docx==1.2.0
requests==2.32.4
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0